DrawfyBot/requirements.txt
pyTelegramBotAPI==4.20.0
pillow==10.3.0
psycopg2-binary==2.9.9
flask==3.0.2
python-dotenv==1.0.0
requests==2.31.0